+++ src/CommerceWidgets.php
@@ -83,6 +83,7 @@ class CommerceWidgets extends Plugin
                $event->types[] = \bymayo\commercewidgets\widgets\SubscriptionPlans::class;
                $event->types[] = \bymayo\commercewidgets\widgets\OrdersRecent::class;
             $event->types[] = \bymayo\commercewidgets\widgets\ConversionRate::class;
+            $event->types[] = \bymayo\commercewidgets\widgets\NewVsReturningCustomers::class;
 
             }
         );